[FEAT]로또 결과 출력

+++ src/main/kotlin/lotto/data/LottoNums.kt
@@ -7,9 +7,9 @@ data class LottoNums(
 )
 
 data class LottoResult(
-    var three : Int = 0,
-    var four : Int = 0 ,
-    var five : Int = 0,
-    var bonus : Int = 0,
-    var six : Int = 0,
+    var three : Long = 0,
+    var four : Long = 0 ,
+    var five : Long = 0,
+    var bonus : Long = 0,
+    var six : Long = 0,
     )
